// Neptune.Web/src/app/shared/generated/model/treatment-bmp-type-detail-dto.ts
export class TreatmentBMPTypeDetailDto { 
    TreatmentBMPTypeID?: number;
    TreatmentBMPTypeName?: string | null;
    TreatmentBMPTypeDescription?: string | null;
    IsAnalyzedInModelingModule?: boolean;
    TreatmentBMPModelingTypeID?: number | null;
    ObservationTypes?: Array<TreatmentBMPTypeObservationTypeDetailDto> | null;
    CustomAttributeTypes?: Array<TreatmentBMPTypeCustomAttributeTypeDetailDto> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}