// Neptune.Web/src/app/shared/generated/model/treatment-bmp-type-upsert-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TreatmentBMPTypeCustomAttributeTypeUpsertDto } from './treatment-bmp-type-custom-attribute-type-upsert-dto';
import { TreatmentBMPTypeObservationTypeUpsertDto } from './treatment-bmp-type-observation-type-upsert-dto';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TreatmentBMPTypeUpsertDto { 
    TreatmentBMPTypeName: string;
    TreatmentBMPTypeDescription?: string | null;
    ObservationTypes?: Array<TreatmentBMPTypeObservationTypeUpsertDto> | null;
    CustomAttributeTypes?: Array<TreatmentBMPTypeCustomAttributeTypeUpsertDto> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TreatmentBMPTypeUpsertDtoForm { 
    TreatmentBMPTypeName: FormControl<string>;
    TreatmentBMPTypeDescription?: FormControl<string>;
    ObservationTypes?: FormControl<Array<TreatmentBMPTypeObservationTypeUpsertDto>>;
    CustomAttributeTypes?: FormControl<Array<TreatmentBMPTypeCustomAttributeTypeUpsertDto>>;
}

export class TreatmentBMPTypeUpsertDtoFormControls { 
    public static TreatmentBMPTypeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
                Validators.minLength(1),
                Validators.maxLength(100),
            ],
        }
    );
    public static TreatmentBMPTypeDescription = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
                Validators.maxLength(1000),
            ],
        }
    );
    public static ObservationTypes = (value: FormControlState<Array<TreatmentBMPTypeObservationTypeUpsertDto>> | Array<TreatmentBMPTypeObservationTypeUpsertDto> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<TreatmentBMPTypeObservationTypeUpsertDto>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomAttributeTypes = (value: FormControlState<Array<TreatmentBMPTypeCustomAttributeTypeUpsertDto>> | Array<TreatmentBMPTypeCustomAttributeTypeUpsertDto> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<TreatmentBMPTypeCustomAttributeTypeUpsertDto>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}