// Neptune.Web/src/app/shared/generated/model/treatment-bmp-type-observation-type-upsert-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TreatmentBMPTypeObservationTypeUpsertDto { 
    TreatmentBMPAssessmentObservationTypeID?: number;
    AssessmentScoreWeight?: number | null;
    DefaultThresholdValue?: number | null;
    DefaultBenchmarkValue?: number | null;
    OverrideAssessmentScoreIfFailing?: boolean;
    SortOrder?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TreatmentBMPTypeObservationTypeUpsertDtoForm { 
    TreatmentBMPAssessmentObservationTypeID?: FormControl<number>;
    AssessmentScoreWeight?: FormControl<number>;
    DefaultThresholdValue?: FormControl<number>;
    DefaultBenchmarkValue?: FormControl<number>;
    OverrideAssessmentScoreIfFailing?: FormControl<boolean>;
    SortOrder?: FormControl<number>;
}

export class TreatmentBMPTypeObservationTypeUpsertDtoFormControls { 
    public static TreatmentBMPAssessmentObservationTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static AssessmentScoreWeight = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static DefaultThresholdValue = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static DefaultBenchmarkValue = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OverrideAssessmentScoreIfFailing = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static SortOrder = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}