// Neptune.Web/src/app/shared/generated/model/treatment-bmp-type-custom-attribute-type-upsert-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TreatmentBMPTypeCustomAttributeTypeUpsertDto { 
    CustomAttributeTypeID?: number;
    SortOrder?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TreatmentBMPTypeCustomAttributeTypeUpsertDtoForm { 
    CustomAttributeTypeID?: FormControl<number>;
    SortOrder?: FormControl<number>;
}

export class TreatmentBMPTypeCustomAttributeTypeUpsertDtoFormControls { 
    public static CustomAttributeTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static SortOrder = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}