// Neptune.Web/src/app/shared/generated/model/treatment-bmp-assessment-observation-type-grid-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TreatmentBMPAssessmentObservationTypeGridDto { 
    TreatmentBMPAssessmentObservationTypeID?: number;
    TreatmentBMPAssessmentObservationTypeName?: string | null;
    ObservationTypeSpecificationID?: number;
    ObservationTypeCollectionMethodDisplayName?: string | null;
    ObservationTargetTypeDisplayName?: string | null;
    ObservationThresholdTypeDisplayName?: string | null;
    TreatmentBMPTypeCount?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TreatmentBMPAssessmentObservationTypeGridDtoForm { 
    TreatmentBMPAssessmentObservationTypeID?: FormControl<number>;
    TreatmentBMPAssessmentObservationTypeName?: FormControl<string>;
    ObservationTypeSpecificationID?: FormControl<number>;
    ObservationTypeCollectionMethodDisplayName?: FormControl<string>;
    ObservationTargetTypeDisplayName?: FormControl<string>;
    ObservationThresholdTypeDisplayName?: FormControl<string>;
    TreatmentBMPTypeCount?: FormControl<number>;
}

export class TreatmentBMPAssessmentObservationTypeGridDtoFormControls { 
    public static TreatmentBMPAssessmentObservationTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static TreatmentBMPAssessmentObservationTypeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ObservationTypeSpecificationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ObservationTypeCollectionMethodDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ObservationTargetTypeDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ObservationThresholdTypeDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static TreatmentBMPTypeCount = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}